Rename callstream pkg to clientstream
Since we're not exposing per-call client options anymore, there's no
`Call` type and the `callstream` name no longer makes sense. This commit
renames the package to `clientstream`.
